Implemented the move editor page
